+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - SC-1379 Validation inputs accepts a react node description
+- SC-1379 In requests, allow to bypass the checkstatus redirection to login page when getting a 401 error
 
 ## 0.0.37